pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reticle"
version = "0.1.0"
description = "Visual agent toolkit — grid overlays and spatial tools for LLM agents"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "numpy>=1.24",
    "pillow>=10.0",
]

[project.optional-dependencies]
openai = ["openai>=1.60"]
anthropic = ["anthropic>=0.40"]
gemini = ["google-genai>=1.0"]
all = ["openai>=1.60", "anthropic>=0.40", "google-genai>=1.0"]
dev = ["ruff>=0.8", "pyright>=1.1"]

[tool.hatch.build.targets.wheel]
packages = ["src/reticle"]

[tool.ruff]
target-version = "py311"
line-length = 99
